requirements.txt
discord.py
requests
pytz
flask
python-dotenv